requirements.txt
pillow
git+https://github.com/Taylor-eOS/last_folder_helper.git
numpy
